ai_status_mailer/requirements.txt
pandas
openpyxl
groq
python-dotenv
